requirements.txt
django>=1.8,<1.9
django-debug-toolbar==1.5
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==3.0

django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
Pillow==3.1.0
